Language: Make `keys` into a single-argument funcntion.
In the new design that emerges from the parser refactoring, an
ambiguity has revealed itself. It turns out that an immediate
(constant or arity-0) filter is a equivalent to a filter function
(arity-1) whose argument is `id`. `jq` chose to make such filters
immediate, probably because calling functions with parentheses makes
using such filter more noisy.

On the other hand, with our syntax and a generally functional approach
it seems more appropriate to add the extra argument in order to make
parsing easier and more consistent with arity-2 infix operators such
as (+) or (*), where `id` filter tends to appear anyways.