Parser: Add operator precedence.
Each operator is now defined as a separate sub-parser (see `operators`
in Parser.Language). These sub-parsers are then combined using the
alternative operator `(<|>)` in the order of decreasing precedence
(notice the list is combined right-to-left). The key to making this
work is parsing the operator and following argument as optional and
returning the whole left argument as a complete expression in case of
failure. Left argument cannot contain the operator in question (though
it might contain operators of higher precedence), so the parser does
not loop on itself, thus always terminating.